WOP: suspense boundry on anime page
