+++ app/radio/loading.tsx
@@ -1,0 +1,3 @@
+export default function AnimeLoading() {
+  return <h1>Loading Animes...</h1>
+}
