+++ app/radio/styles.module.css
@@ -4,15 +4,6 @@
   padding: 0 1em;
 }
 
-.list {
-  display: grid;
-  width: 100%;
-  grid-template-columns: repeat(auto-fit, minmax(235px, 1fr));
-  place-items: center;
-  gap: 1rem;
-  padding: 0;
-  list-style: none;
-}
 .title {
   color: var(--Primary-Color);
   text-align: center;
