+++ components/AnimeList/styles.module.css
@@ -1,0 +1,9 @@
+.list {
+  display: grid;
+  width: 100%;
+  grid-template-columns: repeat(auto-fit, minmax(235px, 1fr));
+  place-items: center;
+  gap: 1rem;
+  padding: 0;
+  list-style: none;
+}
